feat(progress): 支持大数值进度显示和IO进度跟踪
添加对int64大数值进度的支持，包括进度显示格式化和自动宽度计算
新增maxSize和curSize组件用于字节大小显示
添加Write、WrapReader和WrapWriter方法支持IO进度跟踪
更新文档说明新增功能和用法

+++ progress/multi.go
@@ -44,7 +44,7 @@ func (mp *MultiProgress) Add(p *Progress) *Progress {
 }
 
 // New creates a managed progress bar.
-func (mp *MultiProgress) New(maxSteps ...int) *Progress {
+func (mp *MultiProgress) New(maxSteps ...int64) *Progress {
 	return mp.Add(New(maxSteps...))
 }
 
@@ -115,7 +115,7 @@ func (mp *MultiProgress) update(fn func()) {
 	mp.refreshLocked()
 }
 
-func (mp *MultiProgress) startProgress(p *Progress, maxSteps ...int) {
+func (mp *MultiProgress) startProgress(p *Progress, maxSteps ...int64) {
 	mp.mu.Lock()
 	defer mp.mu.Unlock()
